pyproject.toml
[tool.poetry]
name = "bottomless"
version = "0.1.0"
description = ""
authors = ["Jacob Sansbury <sansburyjacob@gmail.com>"] # add yourself here
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.9"
anthropic = "^0.34.1"
pytest-playwright = "^0.5.2"
discord = "^2.3.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
